requirements.txt
schedule
tweepy
groq
python-dotenv